ctrl/requirements.txt
cobs==1.1.3
sseclient-py==1.7
